\name{ELmethodVar}
\alias{ELmethodVar}
\title{Empirical Likelihood Inference of Variance Components in Linear Mixed-Effects Models}
\description{This package provides empirical likelihood-based methods for the inference of  variance components in linear mixed-effects models.}
\author{
    Jingru Zhang, Haochang Shou, Hongzhe Li

    Maintainer: Jingru Zhang (jingru.zhang@pennmedicine.upenn.edu)
}
\references{
    Zhang J., Guo W., Carpenter J.S., Leroux A., Merikangas K.R., Martin N.G., Hickie I.B., Shou H., and Li H. (2022). Empirical likelihood tests for variance components in linear
mixed-effects models. 
}

\seealso{
    \code{\link{ELvar}}
    \code{\link{GELvar}}
}