\name{Philist}
\alias{Philist}
\title{Design Matrices of Variance Components}
\description{This is a list of design matrices of variance components. Its i-th element is an ni by d*ni matrix that combines design matrices of variance components by columns for the i-th subject, where ni is the number of repeated measures for the i-th subject and d is the number of variance components.}